+++ client/src/components/home/HomeIntro.vue
@@ -22,5 +22,30 @@ import SectionLayout from '@/components/layout/SectionLayout.vue'
         <Button icon="pi pi-arrow-right" label="Learn more" rounded />
       </a>
     </div>
+    <div class="flex justify-center mt-4 py-4">
+      <h1 class="font-ham text-xl font-semibold m-auto p-4 dark:text-white/80">
+        <span>How to cite</span>
+      </h1>
+    </div>
+    <div class="flex flex-col item-stretch text-700 line-height-3 -mt-4 dark:text-white/80">
+      <div class="self-center">
+        Etienne Boileau, Harald Wilhelmi, Anne Busch, Andrea Cappannini, Andreas Hildebrand, Janusz
+        M. Bujnicki, Christoph Dieterich.
+      </div>
+      <div class="self-center">
+        <a
+          class="text-primary-500 hover:text-secondary-500"
+          href="https://doi.org/10.1093/nar/gkae972"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Sci-ModoM: a quantitative database of transcriptome-wide high-throughput RNA modification
+          sites
+        </a>
+      </div>
+      <div class="self-center">
+        <span class="italic">Nucleic Acids Research</span>, 2024, gkae972.
+      </div>
+    </div>
   </SectionLayout>
 </template>